Set up Django_VM
I set up another explicit VM for following the Try Django learning resources. This VM does NOT have postgresql installed for now. But it has jupyter installed so that I can document the learning journey and directly interface with the VM environment via a jupyter notebook. I believe documenting the steps - even if a lot of overhead initially - will be worth it in the long run since it makes my work reproducible and I have a self-written reference to return to later.