Barebones LUCE website up and running
Also found the cause for previous error and resolved the nested .git folder problem so commit works again. (The problem was that I had cloned another repo, web3-auth, inside this repo. Github does not allow for nesting of repositories.) Deleting the .git .gitignore and .github folders/files of the sub-directory fixed the issue.