Added new port forwarding & script for debugging
If the Jupyter notebook server is running on the host system (for example via Anaconda) it will 'mask' the port on which LuceVM's Jupyter server is listening. Ideally we should stick with the default port of 8888 but for trouble-shooting we can attempt to run the server over port 8889 instead and see if that fixes the problem.

That said to avoid confusion it is best to ensure only one instance of the Jupyter server is running at any given time.